requirements.txt
Flask==2.3.3
flask-cors==4.0.0
numpy==1.26.3
opencv-python==4.10.0.84
Pillow==10.3.0
tensorflow==2.18.0